all app logic refactored into context
