+++ src/App.jsx
@@ -1,82 +1,14 @@
-import { useState } from "react";
-
 import Header from "./components/Header.jsx";
 import Shop from "./components/Shop.jsx";
 import Product from "./components/Product.jsx";
-import { DUMMY_PRODUCTS } from "./dummy-products.js";
-import { CartContext } from "./store/shopping-cart-context.jsx";
+import CartContextProvider from "./store/shopping-cart-context.jsx";
+import { DUMMY_PRODUCTS } from "./dummy-products";
 
 function App() {
-  const [shoppingCart, setShoppingCart] = useState({
-    items: [],
-  });
-
-  function handleAddItemToCart(id) {
-    setShoppingCart((prevShoppingCart) => {
-      const updatedItems = [...prevShoppingCart.items];
-
-      const existingCartItemIndex = updatedItems.findIndex(
-        (cartItem) => cartItem.id === id
-      );
-      const existingCartItem = updatedItems[existingCartItemIndex];
-
-      if (existingCartItem) {
-        const updatedItem = {
-          ...existingCartItem,
-          quantity: existingCartItem.quantity + 1,
-        };
-        updatedItems[existingCartItemIndex] = updatedItem;
-      } else {
-        const product = DUMMY_PRODUCTS.find((product) => product.id === id);
-        updatedItems.push({
-          id: id,
-          name: product.title,
-          price: product.price,
-          quantity: 1,
-        });
-      }
-
-      return {
-        items: updatedItems,
-      };
-    });
-  }
-
-  function handleUpdateCartItemQuantity(productId, amount) {
-    setShoppingCart((prevShoppingCart) => {
-      const updatedItems = [...prevShoppingCart.items];
-      const updatedItemIndex = updatedItems.findIndex(
-        (item) => item.id === productId
-      );
-
-      const updatedItem = {
-        ...updatedItems[updatedItemIndex],
-      };
-
-      updatedItem.quantity += amount;
-
-      if (updatedItem.quantity <= 0) {
-        updatedItems.splice(updatedItemIndex, 1);
-      } else {
-        updatedItems[updatedItemIndex] = updatedItem;
-      }
-
-      return {
-        items: updatedItems,
-      };
-    });
-  }
-
-  const contextValue = {
-    items: shoppingCart.items, // context items value is the value of our declared state at the beginninf
-    addItemsToCart: handleAddItemToCart, // the add items to cart function can now be stored and passed as a prop
-    updateCart: handleUpdateCartItemQuantity
-  }
-
   return (
     // the app is wrapped in react's createContext function. Provider is a component property created by React
     // same default value set as in the Create Context file. In this case the same as the useState declared at the beginning
-    <CartContext.Provider value={contextValue}>
+    <CartContextProvider>
       <Header />
       <Shop>
         {DUMMY_PRODUCTS.map((product) => (
@@ -85,23 +17,8 @@ function App() {
           </li>
         ))}
       </Shop>
-    </CartContext.Provider>
+    </CartContextProvider>
   );
 }
 
 export default App;
-
-// an alternative to Context Provider is Consumer, which may appear in some older coder bases
-// it uses an an opening an closing tag, in between it dynamically {} outputs a function inside curly braces
-// the function carried out takes the context as a parameter and returns the React components inside the function
-
-{/* <CardContext.Consumer>
-      {(contextValue) => {
-        return (
-          <div>
-            <p>your React code</p>
-          </div>
-        )
-      }
-    }
-    </CardContext.Consumer> */}
